feat(python): Add pybind11 bindings for RasterRenderer and SceneView - Added raster_renderer_py.ipp and scene_view_py.ipp to provide pybind11 bindings for RasterRenderer and SceneView C++ classes. - Updated huira_module.cpp to include and register these new bindings for all spectral specializations. - Exposed the ObservationMode enum to Python. - Updated the star_field.py example to use the new Python bindings for SceneView and RasterRenderer.

+++ bindings/python/include/huira/images/image_py.ipp
@@ -124,7 +124,7 @@ namespace huira {
                     static_cast<py::ssize_t>(img.height()),
                         static_cast<py::ssize_t>(img.width()) });
                 std::memcpy(arr.mutable_data(), img.data(),
-                    static_cast<std::size_t>(img.width()) * img.height()
+                    static_cast<std::size_t>(img.width()) * static_cast<std::size_t>(img.height())
                     * sizeof(Scalar));
                 return arr;
             }
@@ -134,7 +134,7 @@ namespace huira {
                         static_cast<py::ssize_t>(img.width()),
                         static_cast<py::ssize_t>(C) });
                 std::memcpy(arr.mutable_data(), img.data(),
-                    static_cast<std::size_t>(img.width()) * img.height()
+                    static_cast<std::size_t>(img.width()) * static_cast<std::size_t>(img.height())
                     * C * sizeof(Scalar));
                 return arr;
             }
